Fix documentation map function + add example

+++ src/Def/Def.hpp
@@ -17,7 +17,7 @@ constexpr pin_t NO_PIN = 1 << (8 * sizeof(pin_t) - 1);
 
 /// A function pointer to a mapping function to map analog values.
 /// @see    MIDIFilteredAnalog::map()
-using MappingFunction = int (*)(int);
+using MappingFunction = uint8_t (*)(uint8_t);
 
 /// An easy alias for two-dimensional Arrays.
 template <class T, size_t nb_rows, size_t nb_cols>
